services/rag-svc/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.8.2
pydantic-settings==2.4.0
python-dotenv==1.0.1
python-jose==3.3.0
chromadb==0.5.0
openai>=1.3.0
requests==2.32.3
